pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "epoch-tracker"
version = "0.1.0"
description = "Semi-automated system for tracking AI models trained with over 1e25 FLOP"
authors = [{name = "Epoch AI", email = "contact@epoch.ai"}]
readme = "README.md"
requires-python = ">=3.9"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
dependencies = [
    "requests>=2.31.0",
    "pydantic>=2.4.0", 
    "python-dateutil>=2.8.0",
    "huggingface-hub>=0.17.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    "pandas>=2.1.0",
]

[project.optional-dependencies]
scraping = [
    "beautifulsoup4>=4.12.0",
    "selenium>=4.15.0",
]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0", 
    "responses>=0.23.0",
    "black>=23.9.0",
    "isort>=5.12.0",
    "flake8>=6.1.0",
]

[project.scripts]
epoch-scrape = "epoch_tracker.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 88
target-version = ['py39']

[tool.isort]
profile = "black"
multi_line_output = 3